+++ src/components/PlayerView/events.ts
@@ -62,6 +62,7 @@ import {
   VideoPlaybackQualityChangedEvent,
   DownloadFinishedEvent,
   VideoDownloadQualityChangedEvent,
+  PlaybackSpeedChangedEvent,
 } from '../../events';
 
 /**
@@ -244,6 +245,11 @@ interface EventProps {
    * Event emitted when the playback of the current media has finished.
    */
   onPlaybackFinished: PlaybackFinishedEvent;
+  /**
+   * Emitted when the player transitions from one playback speed to another.
+   * @platform iOS, tvOS
+   */
+  onPlaybackSpeedChanged: PlaybackSpeedChangedEvent;
   /**
    * Event emitted when a source is loaded into the player.
    * Seeking and time shifting are allowed as soon as this event is seen.